+++ CHANGELOG.md
@@ -46,4 +46,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   added regression test
 - Removed `retain_graph=True` from gradient penalty computation
 - Vectorised R1/R2 gradient penalty computation
+- Optimized optimizer resets using `zero_grad(set_to_none=True)` in `ACXTrainer`